backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sentence-transformers==2.7.0
huggingface-hub==0.20.3
qdrant-client==1.7.0
google-generativeai==0.3.2
python-dotenv==1.0.0
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
transformers==4.36.0
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
requests==2.31.0
httpx==0.25.2
pandas==2.1.4
openpyxl==3.1.2